Test time parsing separately from business logic
